Adc functions

+++ MiniSumo_2012/delay_ms.c
@@ -2,12 +2,12 @@
  * File Name : delay_ms.c
  * Purpose : For longer delays
  * Creation Date : 26-08-2012
- * Last Modified : tor 30 aug 2012 13:54:17
+ * Last Modified : mån 15 okt 2012 01:58:38
  * Created By : Gabriel Fornaeus, <gf@hax0r.se>
  *
  */
 #include "delay_ms.h"
-void delay_ms(uint8_t ms)
+void delay_ms(uint16_t ms)
 {
 	while (ms)
 	{
